refactor: :recycle: rename file invalid_properties_error > not_properties_error
